split contents logic to different files
