+++ contentmgr/sanity_check.go
@@ -8,12 +8,12 @@ import (
 )
 
 func (cm *ContentManager) HandleSanityCheck(cc cid.Cid, errMsg string) {
-	cm.log.Debugf("running sanity check for cid: %s", cc)
+	cm.log.Warnf("running sanity check for cid: %s", cc)
 
 	// get all contents affected by this missing block on estuary or from shuttles
 	var cnts []util.Content
 	where := "id in (select content from obj_refs where object = (select id from objects where cid = ?))"
-	if err := cm.DB.Find(&cnts, where, cc.Bytes()).Error; err != nil {
+	if err := cm.db.Find(&cnts, where, cc.Bytes()).Error; err != nil {
 		cm.log.Errorf("sanity check failed to get content(s) for cid: %s, err: %w", cc.String(), err)
 		return
 	}
@@ -29,7 +29,7 @@ func (cm *ContentManager) HandleSanityCheck(cc cid.Cid, errMsg string) {
 				ContentID: cnt.ID,
 				ErrMsg:    errMsg,
 			}
-			if err := cm.DB.Clauses(clause.OnConflict{DoNothing: true}).Create(chk).Error; err != nil {
+			if err := cm.db.Clauses(clause.OnConflict{DoNothing: true}).Create(chk).Error; err != nil {
 				cm.log.Errorf("failed to create sanity check mark for content: %d, err: %w", cnt.ID, err)
 				return
 			}
